+++ js/editor-control-comment.js
@@ -1,0 +1,1 @@
+// создание хэштегов и комментария для фото пользователя (ТЗ 2.3, 2.4)
