+++ js/editor-control-effects.js
@@ -1,0 +1,1 @@
+// Наложение эффекта на загружаемое изображение (ТЗ 2.2)
