+++ js/editor-control-scale.js
@@ -1,0 +1,1 @@
+// Изменение масштаба загружаемого изображения (ТЗ 2.1)
