+++ js/open-picture.js
@@ -1,0 +1,1 @@
+// Просмотр изображения загруженного другим пользователем на весь экран (ТЗ 4.4, 4-5)
