+++ js/photo-editor.js
@@ -1,0 +1,1 @@
+// контроль работы редактора, открытие-закрытие модального окна, открытие-закрытие формы редактирования изображения (ТЗ 1.1-1.3)
