+++ js/photo-uploader.js
@@ -1,0 +1,1 @@
+// Отправка данных на сервер (ТЗ 3)
