+++ js/post-filter.js
@@ -1,0 +1,1 @@
+// Фильтрация новостоной ленты (ТЗ 5)
