+++ js/user-post.js
@@ -1,0 +1,1 @@
+// состояные открытого на весь экран изображения (ТЗ 4.6)
